requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
krakenex==2.1.0
pandas==2.0.3
numpy==1.24.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
websockets==12.0
aiohttp==3.9.1 